requirements.txt
python-telegram-bot==21.6
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
pytest==8.3.3


